Create `particle_tracker` module
