+++ plasmapy/simulation/__init__.py
@@ -8,17 +8,10 @@ Module containing plasma simulation tools.
 __all__ = [
     "AbstractSimulation",
     "AbstractTimeDependentSimulation",
-    "ParticleTracker",
 ]
 
 from plasmapy.simulation.abstractions import (
     AbstractNormalizations,
     AbstractSimulation,
     AbstractTimeDependentSimulation,
-)
-from plasmapy.simulation.particle_tracker import (
-    IntervalSaveRoutine,
-    NoParticlesOnGridsTerminationCondition,
-    ParticleTracker,
-    TimeElapsedTerminationCondition,
 )
