+++ plasmapy/simulation/particle_tracker/particle_tracker.py
@@ -3,26 +3,15 @@ Module containing the definition for the general particle tracker.
 """
 
 __all__ = [
-    "AbstractSaveRoutine",
-    "AbstractTerminationCondition",
-    "DoNotSaveSaveRoutine",
-    "IntervalSaveRoutine",
-    "NoParticlesOnGridsTerminationCondition",
     "ParticleTracker",
-    "AllParticlesOffGridTerminationCondition",
-    "TimeElapsedTerminationCondition",
 ]
 
 import collections
 import sys
 import warnings
-from abc import ABC, abstractmethod
 from collections.abc import Iterable
-from pathlib import Path
-from typing import Optional
 
 import astropy.units as u
-import h5py
 import numpy as np
 from numpy.typing import NDArray
 from tqdm import tqdm
@@ -31,389 +20,13 @@ from plasmapy.particles import Particle, particle_input
 from plasmapy.plasma.grids import AbstractGrid
 from plasmapy.plasma.plasma_base import BasePlasma
 from plasmapy.simulation.particle_integrators import boris_push
-
-
-class AbstractTerminationCondition(ABC):
-    """Abstract base class containing the necessary methods for a ParticleTracker termination condition."""
-
-    @property
-    def tracker(self) -> Optional["ParticleTracker"]:
-        """Return the `ParticleTracker` object for this termination condition."""
-        return self._particle_tracker
-
-    @tracker.setter
-    def tracker(self, particle_tracker: "ParticleTracker") -> None:
-        self._particle_tracker: ParticleTracker = particle_tracker
-
-    @property
-    @abstractmethod
-    def require_synchronized_dt(self) -> bool:
-        """Return if this termination condition requires a synchronized time step."""
-        ...
-
-    @property
-    @abstractmethod
-    def progress_description(self) -> str:
-        """Return a small string describing the relevant quantity shown on the meter."""
-        ...
-
-    @property
-    @abstractmethod
-    def units_string(self) -> str:
-        """Return a string representation of the units of `total`."""
-        ...
-
-    @property
-    @abstractmethod
-    def is_finished(self) -> bool:
-        """Return `True` if the simulation has finished."""
-        ...
-
-    @property
-    @abstractmethod
-    def progress(self) -> float:
-        """Return a number representing the progress of the simulation (compared to total).
-        This number represents the numerator of the completion percentage.
-        """
-        ...
-
-    @property
-    @abstractmethod
-    def total(self) -> float:
-        """Return a number representing the total number of steps in a simulation.
-        This number represents the denominator of the completion percentage.
-        """
-        ...
-
-
-class TimeElapsedTerminationCondition(AbstractTerminationCondition):
-    """Termination condition corresponding to the elapsed time of a ParticleTracker."""
-
-    def __init__(self, termination_time: u.Quantity) -> None:
-        self.termination_time: float = termination_time.to(u.s).value
-
-    @property
-    def require_synchronized_dt(self) -> bool:
-        """The elapsed time termination condition requires a synchronized step
-        to stop all particles at the same time.
-        """
-        return True
-
-    @property
-    def progress_description(self) -> str:
-        """The time elapsed termination condition depends on elapsed time,
-        therefore the relevant quantity is time remaining.
-        """
-        return "Time remaining"
-
-    @property
-    def units_string(self) -> str:
-        """The units for the time elapsed condition have the units of seconds."""
-
-        return "seconds"
-
-    @property
-    def is_finished(self) -> bool:
-        """Conclude the simulation if all particles have been tracked over the specified termination time."""
-
-        return bool(float(self.tracker.time) >= self.termination_time)
-
-    @property
-    def progress(self) -> float:
-        """Return the current time step of the simulation."""
-        return float(self.tracker.time)
-
-    @property
-    def total(self) -> float:
-        """Return the total amount of time over which the particles are tracked."""
-        return self.termination_time
-
-
-class NoParticlesOnGridsTerminationCondition(AbstractTerminationCondition):
-    """Termination condition corresponding to stopping the simulation when all particles have exited the grid."""
-
-    def __init__(self) -> None:
-        pass
-
-    @property
-    def require_synchronized_dt(self) -> bool:
-        """The no field termination condition does not require a synchronized time step."""
-        return False
-
-    @property
-    def progress_description(self) -> str:
-        """The progress meter is described in terms of the fraction of particles still on the grid."""
-        return "Number of particles still on grid"
-
-    @property
-    def units_string(self) -> str:
-        """The progress meter is described in terms of the fraction of particles still on the grid."""
-        return "Particles"
-
-    @property
-    def is_finished(self) -> bool:
-        """The simulation is finished when no more particles are on any grids."""
-        is_not_on_grid = self.tracker.on_any_grid == False  # noqa: E712
-
-        return is_not_on_grid.all() and self.tracker.iteration_number > 0
-
-    @property
-    def progress(self) -> float:
-        """The progress of the simulation is measured by how many particles are no longer on a grid."""
-        is_not_on_grid: NDArray[np.bool_] = self.tracker.on_grid == 0
-
-        return float(is_not_on_grid.sum())
-
-    @property
-    def total(self) -> float:
-        """The progress of the simulation is measured against the total number of particles in the simulation."""
-        return float(self.tracker.nparticles)
-
-
-class AllParticlesOffGridTerminationCondition(AbstractTerminationCondition):
-    """Termination condition corresponding to stopping the simulation when a specified
-    proportion of particles have entered and exited the grids.
-
-    Parameters
-    ----------
-    fraction_exited_threshold: float, optional
-        The fraction of particles that must leave the grids to terminate the simulation.
-        This does not include particles that have never entered the grids.
-    """
-
-    def __init__(
-        self,
-        fraction_exited_threshold: float = 0.01,
-    ):
-        super().__init__()
-
-        self.fraction_exited_threshold = fraction_exited_threshold
-
-    @property
-    def require_synchronized_dt(self) -> bool:
-        """The termination condition does not require a synchronized time step."""
-        return False
-
-    @property
-    def progress_description(self) -> str:
-        """The termination condition tracks the number of particles remaining in the grids."""
-        return "Fraction of particles remaining"
-
-    @property
-    def units_string(self) -> str:
-        """The termination condition tracks particles."""
-
-        return "Particles"
-
-    @property
-    def is_finished(self) -> bool:
-        r"""
-        Check to see if the proportion of particles that have entered and exited the grid meet thresholds.
-        """
-
-        # Of the particles that have entered the grid, how many are currently
-        # on the grid?
-        # if/else avoids dividing by zero
-        if self._particle_tracker.num_entered > 0:
-            # Normalize to the number that have entered a grid
-            still_on = (
-                np.sum(self._particle_tracker.on_any_grid)
-                / self._particle_tracker.num_entered
-            )
-        else:
-            still_on = 0.0
-
-        return (
-            self._particle_tracker.num_entered > 0
-            and self.fraction_exited_threshold > still_on
-        )
-
-    @property
-    def progress(self) -> float:
-        """The progress of the simulation is defined in the context of how many particles are on the grids."""
-        return self._particle_tracker.on_any_grid.sum()
-
-    @property
-    def total(self) -> float:
-        """The progress of the simulation is compared to the number of particles tracked."""
-        return self._particle_tracker.nparticles_tracked
-
-
-class AbstractSaveRoutine(ABC):
-    """Abstract base class containing the necessary methods for a
-    |ParticleTracker| save routine.
-
-    The save routine class is responsible for defining the conditions and hooks
-    for saving.
-
-    Parameters
-    ----------
-    output_directory : `~pathlib.Path`, optional
-        Output for objects that are saved to disk. If a directory is not specified
-        then a memory save routine is used.
-
-
-    Notes
-    -----
-    After every push, the `post_push_hook` method is called with the
-    respective `~plasmapy.simulation.particle_tracker.ParticleTracker` object passed as a parameter.
-    Then, the hook calls `save_now` to determine whether or not the simulation state should be saved.
-    """
-
-    def __init__(self, output_directory: Path | None = None) -> None:
-        self.output_directory = output_directory
-
-        self._results = {}
-        self._quantities = {
-            "x": (u.m, "dataset"),
-            "v": (u.m / u.s, "dataset"),
-        }
-
-        self._particle_tracker: ParticleTracker | None = None
-
-    @property
-    def tracker(self) -> Optional["ParticleTracker"]:
-        """Return the `ParticleTracker` object for this stop condition."""
-        return self._particle_tracker
-
-    @tracker.setter
-    def tracker(self, particle_tracker: "ParticleTracker") -> None:
-        self._particle_tracker = particle_tracker
-
-    @property
-    @abstractmethod
-    def require_synchronized_dt(self) -> bool:
-        """Return if this save routine requires a synchronized time step."""
-        ...
-
-    @property
-    @abstractmethod
-    def save_now(self) -> bool:
-        """Determine if to save on the current push step."""
-        ...
-
-    @property
-    def results(self) -> dict[str, u.Quantity]:
-        """Return the results of the simulation.
-        The quantities returned depend on those defined in the body of the save routine.
-        """
-        return self._apply_units_to_results()
-
-    def save(self) -> None:
-        """Save the current state of the simulation to memory.
-
-        If an output directory is specified then the state will also be saved
-        to the disk.
-        """
-
-        self._save_to_memory()
-
-        if self.output_directory is not None:
-            self._save_to_disk()
-
-    def _save_to_disk(self) -> None:
-        """Save a hdf5 file containing simulation positions and velocities."""
-
-        path = self.output_directory / f"{self.tracker.iteration_number}.hdf5"
-
-        with h5py.File(path, "w") as output_file:
-            for key, (_units, data_type) in self._quantities.items():  # noqa: B007
-                match data_type:
-                    case "attribute":
-                        output_file.attrs.create(key, self._results[key])
-                    case "dataset":
-                        output_file.create_dataset(key, data=self._results[key])
-
-    # TODO: Find a better name for this method
-    def _save_to_memory(self) -> None:
-        """Update the results dictionary with the current state of the simulation."""
-
-        for quantity in self._quantities:
-            quantity_history = self._results.get(quantity, [])
-            current_quantity = getattr(self._particle_tracker, quantity, 0)
-
-            quantity_history.append(current_quantity)
-            self._results[quantity] = quantity_history
-
-    def _apply_units_to_results(self):
-        """Apply units to the results dictionary.
-
-        This function is called anytime the results dictionary is retrieved
-        """
-        results_copy = self._results.copy()
-
-        for quantity, (units, _data_type) in self._quantities.items():  # noqa: B007
-            results_copy[quantity] = results_copy[quantity] * units
-
-        return results_copy
-
-    def post_push_hook(self) -> None:
-        """Function called after a push step.
-
-        This function is responsible for handling two steps of save routine, namely:
-            - Deciding to save on the current time step
-            - How the simulation data is saved (i.e. to disk or memory)
-
-        """
-
-        # Update the result dictionary
-        if self.save_now:
-            self.save()
-
-
-class DoNotSaveSaveRoutine(AbstractSaveRoutine):
-    """The default save routine for the |ParticleTracker| class.
-
-    This save routine is a placeholder and will not save the state of the particle tracker.
-    """
-
-    def __init__(self) -> None:
-        super().__init__()
-
-    @property
-    def require_synchronized_dt(self) -> bool:
-        """The do not save save routine does not require a synchronized time step."""
-        return False
-
-    @property
-    def save_now(self) -> bool:
-        """The do not save save routine will never save by definition."""
-        return False
-
-
-class IntervalSaveRoutine(AbstractSaveRoutine):
-    """Abstract class describing a save routine that saves every given interval."""
-
-    def __init__(self, interval: u.Quantity, **kwargs) -> None:
-        super().__init__(**kwargs)
-
-        self._quantities = {
-            "time": (u.s, "dataset"),
-            "x": (u.m, "dataset"),
-            "v": (u.m / u.s, "dataset"),
-        }
-
-        self.save_interval: float = interval.to(u.s).value
-        self.time_of_last_save: float = 0
-
-    @property
-    def require_synchronized_dt(self) -> bool:
-        """Save output only makes sense for synchronized time steps."""
-        return True
-
-    @property
-    def save_now(self) -> bool:
-        """Save at every interval given in instantiation."""
-
-        return bool(self.tracker.time - self.time_of_last_save >= self.save_interval)
-
-    def save(self) -> None:
-        """Save the current state of the simulation.
-        Sets the time of last save attribute and log the timestamp.
-        """
-        super().save()
-
-        self.time_of_last_save = self.tracker.time
+from plasmapy.simulation.particle_tracker.save_routines import (
+    AbstractSaveRoutine,
+    DoNotSaveSaveRoutine,
+)
+from plasmapy.simulation.particle_tracker.termination_conditions import (
+    AbstractTerminationCondition,
+)
 
 
 class ParticleTracker:
